requirements.txt
feedparser==6.0.10
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
